requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
PyPDF2>=3.0.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.1.0
tqdm>=4.66.0
